Moving base methods from AppMgtRestClient to BaseRestClient
